# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface EmailDomain {
  name: string;
}

export interface GenerateEmailRequest {
  /** Optional domain to use for the email address */
  domain?: string;
}

export interface EmailAccount {
  id: string;
  address: string;
  token: string;
  createdAt: string;
  expiresAt: string;
}

export interface EmailMessage {
  id: string;
  from: string;
  subject: string;
  intro: string;
  receivedAt: string;
  hasOtp: boolean;
}

export interface EmailMessageFull {
  id: string;
  from: string;
  subject: string;
  body: string;
  receivedAt: string;
  /** @nullable */
  otp: string | null;
}

export interface OtpResult {
  /** @nullable */
  otp: string | null;
  /** @nullable */
  source: string | null;
  /** @nullable */
  foundIn: string | null;
}