# lib/api-zod/src/generated/types/emailAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailAccount {
  id: string;
  address: string;
  token: string;
  createdAt: string;
  expiresAt: string;
}