# lib/api-zod/src/generated/types/emailMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailMessage {
  id: string;
  from: string;
  subject: string;
  intro: string;
  receivedAt: string;
  hasOtp: boolean;
}