# lib/api-zod/src/generated/types/emailMessageFull.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailMessageFull {
  id: string;
  from: string;
  subject: string;
  body: string;
  receivedAt: string;
  /** @nullable */
  otp: string | null;
}