# lib/api-zod/src/generated/types/generateEmailRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateEmailRequest {
  /** Optional domain to use for the email address */
  domain?: string;
}