# lib/api-zod/src/generated/types/otpResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OTP Email Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OtpResult {
  /** @nullable */
  otp: string | null;
  /** @nullable */
  source: string | null;
  /** @nullable */
  foundIn: string | null;
}